requirements.txt
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
et_xmlfile==2.0.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
numpy==2.3.3
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.2.4
pypdf==6.0.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
